Updated, kept all the old equations but described why we are omitted PE for non significant values
Since the cross PE of Tide64 does not effect Tide 128, ran the same function but omitted it. Same with Tide64 revenue in relation to Wisk64, seems that Wisk64 cross PE does not have a significant effect on Tide64.  The final table is the last thing that needs to be updated, because the outputs were slightly different.